pyproject.toml
[project]
name = "insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.21.0",
    "langchain-openai>=1.1.6",
    "litellm>=1.80.11",
    "mysql>=0.0.3",
    "mysql-connector-python==8.4",
    "pandas>=2.3.3",
    "pydantic-ai>=1.31.0",
    "pymysql>=1.1.2",
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
